+++ src/MapperQuery/OptionsQueryParam.php
@@ -11,6 +11,25 @@ class OptionsQueryParam extends AbstractQueryParam
 {
     private ?array $cleanedParts = null;
 
+    public function setQueryParamString(string $value): static
+    {
+        $this->cleanedParts = null;
+        return parent::setQueryParamString($value);
+    }
+
+    public function addQueryParamString(string $value): static
+    {
+        if ($this->getQueryParamString === '') {
+            $this->setQueryParamString($value);
+        } else {
+            $this->setQueryParamString(
+                $this->queryParamString . ',' . $value
+            );
+        }
+
+        return $this;
+    }
+
     public function getParts(): array
     {
         return $this->cleanedParts ?? parent::getParts();
@@ -34,7 +53,9 @@ class OptionsQueryParam extends AbstractQueryParam
 
         $newParts = [];
 
-        foreach ($this->getParts() as $value) {
+        $parts = array_unique($this->getParts());
+
+        foreach ($parts as $value) {
             if (call_user_func($validate, $value)) {
                 $newParts[] = $value;
             }