+++ src/MapperQuery/OrderByQueryParam.php
@@ -13,6 +13,25 @@ class OrderByQueryParam extends AbstractQueryParam
 {
     private ?array $cleanedParts = null;
 
+    public function setQueryParamString(string $value): static
+    {
+        $this->cleanedParts = null;
+        return parent::setQueryParamString($value);
+    }
+
+    public function addQueryParamString(string $value): static
+    {
+        if ($this->getQueryParamString === '') {
+            $this->setQueryParamString($value);
+        } else {
+            $this->setQueryParamString(
+                $this->queryParamString . ',' . $value
+            );
+        }
+
+        return $this;
+    }
+
     public function getParts(): array
     {
         return $this->cleanedParts ?? parent::getParts();